Cleanup and Fix CSV Formatting
Cleaned up a few items, and I made changes to fix the CSV, so it's formatted correctly. I also added a small bit of logging, so you can watch your scripts progress. 

It appears there are some Agency Names missing, I'm not sure if that's because they aren't there or because our scraping sometimes fails, probably want to look into that and add some error checking if applicable.

Take a look at my comments for details on the changes, then hit me up on Slack with any questions!

-Will